refactor(app): reorganize store structure by separating game and answers slices into distinct files; update imports for clarity; where possible, I used interface instead of type; replaced the function declaration with an arrow function
