+++ src/app/providers/theme/theme-provider.types.ts
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import type { ReactNode } from 'react';
 
 export interface CustomThemeProviderProps {
   children: ReactNode;
