+++ src/app/store/answers/answers.types.ts
@@ -1,0 +1,14 @@
+interface CorrectAnswers {
+  [indexQuestion: number]: string[];
+}
+
+export interface AnswersSlice {
+  checkedAnswers: string[];
+  numberOfAnswers: number;
+  userAnswers: CorrectAnswers;
+  clearCheckedAnswers: () => void;
+  toggleAnswer: (userAnswer: string) => void;
+  increaseNumberOfAnswers: () => void;
+  updateUserAnswers: (answers: string[], indexQuestion: number) => void;
+  clearNumberOfAnswers: () => void;
+}
